satisfy rubocop
also unify on variables and methods using klass instead of class
