+++ lib/interactify/job_maker.rb
@@ -3,6 +3,8 @@
 require "sidekiq"
 require "sidekiq/job"
 
+require "interactify/async_job_klass"
+
 module Interactify
   class JobMaker
     attr_reader :opts, :method_name, :container_klass, :klass_suffix
@@ -15,26 +17,26 @@ module Interactify
     end
 
     concerning :JobClass do
-      def job_class
-        @job_class ||= define_job_class
+      def job_klass
+        @job_klass ||= define_job_klass
       end
 
       private
 
-      def define_job_class
+      def define_job_klass
         this = self
 
         invalid_keys = this.opts.symbolize_keys.keys - %i[queue retry dead backtrace pool tags]
 
         raise ArgumentError, "Invalid keys: #{invalid_keys}" if invalid_keys.any?
 
-        job_class(opts).tap do |klass|
+        build_job_klass(opts).tap do |klass|
           klass.const_set(:JOBABLE_OPTS, opts)
           klass.const_set(:JOBABLE_METHOD_NAME, method_name)
         end
       end
 
-      def job_class(opts)
+      def build_job_klass(opts)
         Class.new do
           include Sidekiq::Job
 
@@ -47,55 +49,8 @@ module Interactify
       end
     end
 
-    concerning :AsyncJobClass do
-      def async_job_class
-        klass = Class.new do
-          include Interactor
-          include Interactor::Contracts
-        end
-
-        attach_call(klass)
-        attach_call!(klass)
-
-        klass
-      end
-
-      def args(context)
-        args = context.to_h.stringify_keys
-
-        return args unless container_klass.respond_to?(:expected_keys)
-
-        restrict_to_optional_or_keys_from_contract(args)
-      end
-
-      def attach_call(async_job_class)
-        # e.g. SomeInteractor::AsyncWithSuffix.call(foo: 'bar')
-        async_job_class.send(:define_singleton_method, :call) do |context|
-          call!(context)
-        end
-      end
-
-      def attach_call!(async_job_class)
-        this = self
-
-        # e.g. SomeInteractor::AsyncWithSuffix.call!(foo: 'bar')
-        async_job_class.send(:define_singleton_method, :call!) do |context|
-          # e.g. SomeInteractor::JobWithSuffix
-          job_klass = this.container_klass.const_get("Job#{this.klass_suffix}")
-
-          # e.g. SomeInteractor::JobWithSuffix.perform_async({foo: 'bar'})
-          job_klass.perform_async(this.args(context))
-        end
-      end
-
-      def restrict_to_optional_or_keys_from_contract(args)
-        keys = container_klass.expected_keys.map(&:to_s)
-
-        optional = Array(container_klass.optional_attrs).map(&:to_s)
-        keys += optional
-
-        args.slice(*keys)
-      end
+    def async_job_klass
+      AsyncJobKlass.new(container_klass:, klass_suffix:).async_job_klass
     end
   end
 end
