+++ spec/spec_helper.rb
@@ -3,7 +3,6 @@
 require "interactify"
 require "deep_matching"
 require "rails"
-require "debug"
 require "sidekiq/testing"
 require "debug"
